refactor(accessible-button): to typescript
